Write failure state tests for the note registry
