+++ packages/protocol/contracts/ACE/validators/adjustSupply/AdjustSupply.sol
@@ -1,23 +1,8 @@
 pragma solidity >=0.5.0 <0.6.0;
 
-import "../../../libs/LibEIP712.sol";
 import "./AdjustSupplyABIEncoder.sol";
-
-contract AdjustSupplyInterface is LibEIP712 {
-    /* solhint-disable-next-line var-name-mixedcase */
-
-    constructor() public {}
-
-    function validateAdjustSupply(
-        bytes calldata, // proof data
-        address, // sender address
-        uint[6] calldata // common reference string
-    ) 
-        external 
-        pure 
-        returns (bytes memory) // returns a series of transfer instructions
-    {}
-}
+import "../../../interfaces/AdjustSupplyInterface.sol";
+import "../../../libs/LibEIP712.sol";
 
 /**
  * @title Library to validate AZTEC zero-knowledge proofs
